Move the perf utils in a local folder
