+++ packages/e2e-tests/specs/performance/utils/events.js
@@ -1,20 +1,3 @@
-/**
- * External dependencies
- */
-import { existsSync, readFileSync, unlinkSync } from 'fs';
-
-export function readFile( filePath ) {
-	return existsSync( filePath )
-		? readFileSync( filePath, 'utf8' ).trim()
-		: '';
-}
-
-export function deleteFile( filePath ) {
-	if ( existsSync( filePath ) ) {
-		unlinkSync( filePath );
-	}
-}
-
 function isKeyEvent( item ) {
 	return (
 		item.cat === 'devtools.timeline' &&
