+++ packages/e2e-tests/specs/performance/utils/files.js
@@ -1,0 +1,16 @@
+/**
+ * External dependencies
+ */
+import { existsSync, readFileSync, unlinkSync } from 'fs';
+
+export function readFile( filePath ) {
+	return existsSync( filePath )
+		? readFileSync( filePath, 'utf8' ).trim()
+		: '';
+}
+
+export function deleteFile( filePath ) {
+	if ( existsSync( filePath ) ) {
+		unlinkSync( filePath );
+	}
+}
