+++ packages/e2e-tests/specs/performance/utils/index.js
@@ -1,0 +1,7 @@
+export {
+	getClickEventDurations,
+	getHoverEventDurations,
+	getSelectionEventDurations,
+	getTypingEventDurations,
+} from './events';
+export { deleteFile, readFile } from './files';
